// types.ts
export interface Pillar {
    icon: string;
    title: string;
    description: string;
    color: string;
}

export interface Product {
    icon: string;
    category: string;
    categoryColor: string;
    title: string;
    description: string;
}

export interface Job {
    icon: string;
    title: string;
    details: string;
    color: string;
}